cos, sin to float32

+++ src/transformers/models/llama/modeling_llama.py
@@ -125,11 +125,14 @@ class LlamaRotaryEmbedding(nn.Module):
         inv_freq_expanded = self.inv_freq[None, :, None].float().expand(position_ids.shape[0], -1, 1)
         position_ids_expanded = position_ids[:, None, :].float()
         # Force float32 since bfloat16 loses precision on long contexts
+        # See https://github.com/huggingface/transformers/pull/29285
         with torch.autocast(device_type=position_ids_expanded.device.type, enabled=False):
             freqs = (inv_freq_expanded.float() @ position_ids_expanded.float()).transpose(1, 2)
-        emb = torch.cat((freqs, freqs), dim=-1)
-        cos = emb.cos().to(dtype=x.dtype)
-        sin = emb.sin().to(dtype=x.dtype)
+            emb = torch.cat((freqs, freqs), dim=-1)
+            cos = emb.cos()
+            sin = emb.sin()
+        cos = cos.to(dtype=x.dtype)
+        sin = sin.to(dtype=x.dtype)
         # backwards compatibility
         self._cos_cached = cos
         self._sin_cached = sin
